backend/requirements.txt
alembic==1.16.5
black==26.3.1
blinker==1.9.0
bootstrap-flask==2.5.0
cfgv==3.5.0
click==8.3.1
coverage==7.13.4
distlib==0.4.0
dnspython==2.7.0
dominate==2.9.1
email-validator==2.3.0
filelock==3.20.3
flake8==7.3.0
flask==3.1.3
flask-cors==6.0.0
flask-login==0.6.3
flask-mail==0.10.0
flask-migrate==4.1.0
flask-moment==1.0.6
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.3.2
gunicorn==23.0.0
identify==2.6.16
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
jinja2==3.1.6
librt==0.8.0
mako==1.3.10
markupsafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy-extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
pre-commit==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pygments==2.20.0
pyjwt==2.12.0
pymysql==1.1.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dotenv==1.2.1
pytokens==0.4.1
pyyaml==6.0.3
sqlalchemy==2.0.46
tomli==2.4.0
typing-extensions==4.15.0
virtualenv==20.36.1
visitor==0.1.3
werkzeug==3.1.6
wtforms==3.2.1
zipp==3.23.0
